requirements.txt
telethon
opencv-python
pytesseract
selenium
python-dotenv